requirements.txt
Django~=4.1.1
psycopg2-binary
Pillow
djangorestframework~=3.14.0
drf-yasg
Django-filter~=22.1
python-decouple
djangorestframework-simplejwt